Fix: Correction des erreurs de base de données et types
- Ajout des tables payment_transactions et token_transactions dans la migration principale
- Correction de l'import Utc dans alert_service.rs
- Correction des types i32/i64 dans payment_service.rs
- Résolution des erreurs E0433 et E0308

+++ backend/src/services/payment_service.rs
@@ -300,7 +300,7 @@ impl PaymentService {
         // Mettre à jour le solde de l'utilisateur
         sqlx::query!(
             "UPDATE users SET tokens_balance = tokens_balance + $1 WHERE id = $2",
-            total_tokens,
+            total_tokens as i64,
             user_id
         )
         .execute(&self.pool)
@@ -315,9 +315,9 @@ impl PaymentService {
             VALUES ($1, $2, $3, $4, $5, $6, NOW())
             "#,
             user_id,
-            tokens,
-            bonus,
-            total_tokens,
+            tokens as i32,
+            bonus as i32,
+            total_tokens as i32,
             "recharge",
             format!("Recharge de {} XAF", amount)
         )
